@@ -14,7 +14,7 @@ namespace ColourAPI.Controllers
         [HttpGet]
         public ActionResult<IEnumerable<string>> Get()
         {
-            return new string[] { "value1", "value2" };
+            return new string[] { "value1", "value2", "Red" };
         }
 
         // GET api/values/5
